Reduce the offsets in DwarfDebugInfoEntry to 32 bit, they're printed with %x and that breaks on big-endian machines.
I have to clean up the 32/64 bit confusion in libDebugInfo some day.

llvm-svn: 143812

+++ llvm/lib/DebugInfo/DWARFDebugInfoEntry.cpp
@@ -26,7 +26,7 @@ void DWARFDebugInfoEntryMinimal::dump(raw_ostream &OS,
   uint32_t offset = Offset;
 
   if (debug_info_data.isValidOffset(offset)) {
-    uint64_t abbrCode = debug_info_data.getULEB128(&offset);
+    uint32_t abbrCode = debug_info_data.getULEB128(&offset);
 
     OS << format("\n0x%8.8x: ", Offset);
     if (abbrCode) {
